refactor(incarnon): address PR review — codec bug, lazy data, stable id
Fixes the round-trip bug, moves the 70KB evolution data out of the web
bundle to a lazy-fetched JSON, and replaces the regex-based attack-mode
filter with a pinned identifier.

- **Codec round-trip**: encoder now emits `ic` whenever
  `incarnonEnabled !== undefined`, not only when truthy. Previously,
  explicitly turning the toggle OFF on an innate incarnon (Phenmor,
  Felarx, Praedos, …) round-tripped to undefined, and the decoder's
  default-on heuristic flipped it back on. Same bug bit saved builds.
- **Static data convention**: split the 70KB perk-tree dataset out of
  `packages/shared` and into `apps/web/public/data/incarnon-evolutions.json`,
  emitted by `build:items`. The web bundle now ships ~3KB (names + genesis
  image map + variant stripper) instead of ~70KB. Editor sidebar fetches
  the full data via a TanStack query (`incarnonEvolutionsQuery`) only when
  the picker actually renders.
- **Cross-file sanity check**: `build:items` throws if `INCARNON_NAMES`
  drifts from `INCARNON_EVOLUTIONS` keys (or genesis image mismatches),
  so the small/large source files can't silently desync.
- **Stable attack-mode identifier**: `INCARNON_FORM_ATTACK_NAME` constant
  replaces the `/incarnon/i.test(m.name)` regex in the stats-panel filter.
  Easier to update if WFCD ever renames the mode.
- **Removed double computation** in `create.tsx`: a single
  `getIncarnonGenesisImage(item.name)` lookup feeds both the auto-toggle
  default and the image swap.
- **Verified `&` in URL**: Ack & Brunt's genesis adapter image
  (`ack-&-brunt-incarnon-genesis-…png`) loads fine via `<img>` despite
  the unencoded ampersand — `getImageUrl` doesn't need encoding for path
  segments and CDN handles it.

No behavior changes for users beyond the codec bug fix.
